+++ app/Contact.php
@@ -377,6 +377,16 @@ class Contact extends Model
         );
     }
 
+    public function getSyndicationUrl()
+    {
+        return \Movim\Route::urlize(
+            'blog',
+            ($this->user && isset($this->user->nickname))
+                ? $this->user->nickname
+                : $this->id
+        );
+    }
+
     public function isBlocked(): bool
     {
         return \App\User::me()->hasBlocked($this->id, true);