+++ CHANGELOG.md
@@ -14,6 +14,8 @@ v0.30.1 (master)
 * Allow to search Channels in the global directory directly from the Search bar
 * Introducing the URL Resolver Worker, that resolves the shared URLs in a non-blocking and concurrential way
 * Refactor the Ad-Hoc widget and move it into the contacts and MUCs drawers
+* Fix #1437 Generalize the instance nickname usage on the public blog and syndication feed, add a warning message if the instance nickname is changed
+* Rename the config page to configuration
 
 v0.30
 ---------------------------